refactor(immutable): move selectImmutableAssets to its own file
