+++ internal/apis/kfd/v1alpha2/immutable/create/immutable_assets.go
@@ -1,0 +1,44 @@
+// Copyright (c) 2017-present SIGHUP s.r.l All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package create
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"gopkg.in/yaml.v2"
+)
+
+// selectImmutableAssets loads the vendored immutable.yaml and selects the block for the given
+// kubernetes version. It is the single selector used by every furyctl phase (infrastructure and
+// kubernetes), so baked artifacts and rendered role variables never disagree on version. A free
+// function (no Infrastructure receiver) so both phases call it directly. The version is passed as-is
+// (it comes from the kfd manifest, which we control), so a mismatch surfaces as
+// ErrKubernetesVersionNotFound; an empty manifest surfaces as ErrNoKubernetesVersions.
+func selectImmutableAssets(phasePath, kubeVersion string) (assets, error) {
+	immutableSpecPath := filepath.Join(phasePath, "..", "vendor", "installers", "immutable", "immutable.yaml")
+
+	data, err := os.ReadFile(immutableSpecPath)
+	if err != nil {
+		return assets{}, fmt.Errorf("error reading immutable manifest at %s: %w", immutableSpecPath, err)
+	}
+
+	var manifest immutableManifest
+	if err := yaml.Unmarshal(data, &manifest); err != nil {
+		return assets{}, fmt.Errorf("error parsing immutable manifest: %w", err)
+	}
+
+	if len(manifest.Kubernetes) == 0 {
+		return assets{}, ErrNoKubernetesVersions
+	}
+
+	immutableAssets, ok := manifest.Kubernetes[kubeVersion]
+	if !ok {
+		return assets{}, fmt.Errorf("%w: %s", ErrKubernetesVersionNotFound, kubeVersion)
+	}
+
+	return immutableAssets, nil
+}
